# [LAB] API REST/assets/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "lab-api-rest",
      "version": "1.0.0",
      "license": {
        "name": "ISC"
      },
      "description": "API REST pour gérer des utilisateurs"
    },
    "basePath":"/api/v1",
    "host":"localhost:8080",
    "tags": [
        {
        "name":"members",
        "description": "Méthodes permettant de gérer les utilisateurs"
        }
    ],
    "paths": {
      "members/{id}": {
        "get": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        },
        "put": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        },
        "delete": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        }
      },
      "members": {
        "get": {
          "summary": "members",
          "parameters": [
            {
              "name": "max",
              "in": "query",
              "required": false,
              "schema": {
                "type": "int"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "result": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        },
                        "example": [
                          {
                            "id": 1,
                            "name": "Yann"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "tags": ["members"]
        },
        "post": {
          "summary": "members",
          "parameters": [
            {
              "name": "name",
              "in": "body",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "result": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "Yann"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "tags": ["members"]
        }
      }
    }
  }